+++ SilKit/source/config/YamlConversion.hpp
@@ -80,6 +80,9 @@ DEFINE_SILKIT_CONVERT(Middleware);
 
 DEFINE_SILKIT_CONVERT(Extensions);
 
+DEFINE_SILKIT_CONVERT(Experimental);
+DEFINE_SILKIT_CONVERT(TimeSynchronization);
+
 DEFINE_SILKIT_CONVERT(ParticipantConfiguration);
 
 } // namespace YAML