+++ SilKit/source/core/participant/Participant_impl.hpp
@@ -158,7 +158,8 @@ void Participant<SilKitConnectionT>::SetupRemoteLogging()
             Config::InternalController config;
             config.name = "LogMsgReceiver";
             config.network = "default";
-            CreateController<Services::Logging::LogMsgReceiver>(config, std::move(supplementalData), true, logger);
+            CreateController<Services::Logging::LogMsgReceiver>(config, std::move(supplementalData), true, true,
+                                                                logger);
         }
 
         auto sinkIter = std::find_if(_participantConfig.logging.sinks.begin(), _participantConfig.logging.sinks.end(),
@@ -174,11 +175,10 @@ void Participant<SilKitConnectionT>::SetupRemoteLogging()
             config.name = "LogMsgSender";
             config.network = "default";
             auto&& logMsgSender =
-                CreateController<Services::Logging::LogMsgSender>(config, std::move(supplementalData), true);
+                CreateController<Services::Logging::LogMsgSender>(config, std::move(supplementalData), true, true);
 
-            logger->RegisterRemoteLogging([logMsgSender](Services::Logging::LogMsg logMsg) {
-                logMsgSender->SendLogMsg(std::move(logMsg));
-            });
+            logger->RegisterRemoteLogging(
+                [logMsgSender](Services::Logging::LogMsg logMsg) { logMsgSender->SendLogMsg(std::move(logMsg)); });
         }
     }
     else
@@ -255,7 +255,7 @@ auto Participant<SilKitConnectionT>::CreateCanController(const std::string& cano
     Core::SupplementalData supplementalData;
     supplementalData[SilKit::Core::Discovery::controllerType] = SilKit::Core::Discovery::controllerTypeCan;
 
-    auto controller = CreateController<Can::CanController>(controllerConfig, std::move(supplementalData), true,
+    auto controller = CreateController<Can::CanController>(controllerConfig, std::move(supplementalData), true, true,
                                                            controllerConfig, &_timeProvider);
 
     controller->RegisterServiceDiscovery();
@@ -291,7 +291,7 @@ auto Participant<SilKitConnectionT>::CreateEthernetController(
     supplementalData[SilKit::Core::Discovery::controllerType] = SilKit::Core::Discovery::controllerTypeEthernet;
 
     auto* controller = CreateController<Ethernet::EthController>(controllerConfig, std::move(supplementalData), true,
-                                                                 controllerConfig, &_timeProvider);
+                                                                 true, controllerConfig, &_timeProvider);
 
     controller->RegisterServiceDiscovery();
 
@@ -326,7 +326,7 @@ auto Participant<SilKitConnectionT>::CreateFlexrayController(
     supplementalData[SilKit::Core::Discovery::controllerType] = SilKit::Core::Discovery::controllerTypeFlexray;
 
     auto controller = CreateController<Flexray::FlexrayController>(controllerConfig, std::move(supplementalData), true,
-                                                                   controllerConfig, &_timeProvider);
+                                                                   true, controllerConfig, &_timeProvider);
 
     controller->RegisterServiceDiscovery();
 
@@ -354,7 +354,7 @@ auto Participant<SilKitConnectionT>::CreateLinController(const std::string& cano
     Core::SupplementalData supplementalData;
     supplementalData[SilKit::Core::Discovery::controllerType] = SilKit::Core::Discovery::controllerTypeLin;
 
-    auto controller = CreateController<Lin::LinController>(controllerConfig, std::move(supplementalData), true,
+    auto controller = CreateController<Lin::LinController>(controllerConfig, std::move(supplementalData), true, true,
                                                            controllerConfig, &_timeProvider);
 
     controller->RegisterServiceDiscovery();
@@ -402,8 +402,8 @@ auto Participant<SilKitConnectionT>::CreateDataSubscriberInternal(
     std::string network = linkName;
 
     auto controller = CreateController<PubSub::DataSubscriberInternal>(
-        controllerConfig, network, std::move(supplementalData), true, &_timeProvider, topic, mediaType, publisherLabels,
-        defaultHandler, parent);
+        controllerConfig, network, std::move(supplementalData), true, true, &_timeProvider, topic, mediaType,
+        publisherLabels, defaultHandler, parent);
 
     //Restore original DataSubscriber config for replay
     auto&& parentConfig = parentDataSubscriber->GetConfig();
@@ -490,8 +490,8 @@ auto Participant<SilKitConnectionT>::CreateDataPublisher(const std::string& cano
     supplementalData[SilKit::Core::Discovery::supplKeyDataPublisherPubLabels] = labelStr;
 
     auto controller = CreateController<Services::PubSub::DataPublisher>(
-        controllerConfig, network, std::move(supplementalData), true, &_timeProvider, configuredDataNodeSpec, network,
-        controllerConfig);
+        controllerConfig, network, std::move(supplementalData), true, true, &_timeProvider, configuredDataNodeSpec,
+        network, controllerConfig);
 
     _connection.SetHistoryLengthForLink(history, controller);
 
@@ -552,7 +552,7 @@ auto Participant<SilKitConnectionT>::CreateDataSubscriber(
     supplementalData[SilKit::Core::Discovery::supplKeyDataSubscriberSubLabels] = labelStr;
 
     auto controller = CreateController<Services::PubSub::DataSubscriber>(
-        controllerConfig, network, std::move(supplementalData), true, controllerConfig, &_timeProvider,
+        controllerConfig, network, std::move(supplementalData), true, true, controllerConfig, &_timeProvider,
         configuredDataNodeSpec, defaultDataHandler);
 
     controller->RegisterServiceDiscovery();
@@ -602,7 +602,7 @@ auto Participant<SilKitConnectionT>::CreateRpcServerInternal(
             std::to_string(parentRpcServer->GetServiceDescriptor().GetServiceId());
     }
     return CreateController<Services::Rpc::RpcServerInternal>(controllerConfig, network, std::move(supplementalData),
-                                                              true, &_timeProvider, functionName, mediaType,
+                                                              true, true, &_timeProvider, functionName, mediaType,
                                                               clientLabels, clientUUID, handler, parent);
 }
 
@@ -635,7 +635,7 @@ auto Participant<SilKitConnectionT>::CreateRpcClient(
     }
 
     auto controller =
-        CreateController<Services::Rpc::RpcClient>(controllerConfig, network, std::move(supplementalData), true,
+        CreateController<Services::Rpc::RpcClient>(controllerConfig, network, std::move(supplementalData), true, true,
                                                    &_timeProvider, configuredDataSpec, network, handler);
 
     // RpcClient discovers RpcServerInternal and is ready to dispatch calls
@@ -683,7 +683,7 @@ auto Participant<SilKitConnectionT>::CreateRpcServer(
     }
 
     auto controller = CreateController<Services::Rpc::RpcServer>(controllerConfig, network, supplementalData, true,
-                                                                 &_timeProvider, configuredDataSpec, handler);
+                                                                 true, &_timeProvider, configuredDataSpec, handler);
 
     // RpcServer discovers RpcClient and creates RpcServerInternal on a matching connection
     controller->RegisterServiceDiscovery();
@@ -713,6 +713,7 @@ auto Participant<SilKitConnectionT>::CreateTimeSyncService(Orchestration::Lifecy
         throw SilKitError("Tried to instantiate TimeSyncService multiple times!");
     }
 
+
     Core::SupplementalData timeSyncSupplementalData;
     timeSyncSupplementalData[SilKit::Core::Discovery::controllerType] =
         SilKit::Core::Discovery::controllerTypeTimeSyncService;
@@ -721,8 +722,8 @@ auto Participant<SilKitConnectionT>::CreateTimeSyncService(Orchestration::Lifecy
     config.name = Discovery::controllerTypeTimeSyncService;
     config.network = "default";
     timeSyncService = CreateController<Orchestration::TimeSyncService>(
-        config, std::move(timeSyncSupplementalData), false, &_timeProvider, _participantConfig.healthCheck,
-        lifecycleService);
+        config, std::move(timeSyncSupplementalData), false, false, &_timeProvider, _participantConfig.healthCheck,
+        lifecycleService, _participantConfig.experimental.timeSynchronization.animationFactor);
 
     return timeSyncService;
 }
@@ -742,8 +743,8 @@ auto Participant<SilKitConnectionT>::GetLifecycleService() -> Services::Orchestr
         Config::InternalController config;
         config.name = Discovery::controllerTypeLifecycleService;
         config.network = "default";
-        lifecycleService =
-            CreateController<Orchestration::LifecycleService>(config, std::move(lifecycleSupplementalData), false);
+        lifecycleService = CreateController<Orchestration::LifecycleService>(
+            config, std::move(lifecycleSupplementalData), false, true);
     }
     return lifecycleService;
 }
@@ -817,7 +818,7 @@ auto Participant<SilKitConnectionT>::GetSystemMonitor() -> Services::Orchestrati
         Config::InternalController config;
         config.name = Discovery::controllerTypeSystemMonitor;
         config.network = "default";
-        controller = CreateController<Orchestration::SystemMonitor>(config, std::move(supplementalData), true);
+        controller = CreateController<Orchestration::SystemMonitor>(config, std::move(supplementalData), true, true);
 
         _connection.RegisterMessageReceiver([controller](IVAsioPeer* peer, const ParticipantAnnouncement&) {
             controller->OnParticipantConnected(
@@ -858,7 +859,7 @@ auto Participant<SilKitConnectionT>::GetServiceDiscovery() -> Discovery::IServic
         config.name = Discovery::controllerTypeServiceDiscovery;
         config.network = "default";
         controller = CreateController<SilKit::Core::Discovery::ServiceDiscovery>(config, std::move(supplementalData),
-                                                                                 true, GetParticipantName());
+                                                                                 true, true, GetParticipantName());
 
         _connection.RegisterPeerShutdownCallback(
             [controller](IVAsioPeer* peer) { controller->OnParticpantRemoval(peer->GetInfo().participantName); });
@@ -887,7 +888,7 @@ auto Participant<SilKitConnectionT>::GetRequestReplyService() -> RequestReply::I
         config.name = "RequestReplyService";
         config.network = "default";
         controller = CreateController<RequestReply::RequestReplyService>(config, std::move(supplementalData), true,
-                                                                         GetParticipantName(), procedures);
+                                                                         true, GetParticipantName(), procedures);
 
         _connection.RegisterPeerShutdownCallback(
             [controller](IVAsioPeer* peer) { controller->OnParticpantRemoval(peer->GetInfo().participantName); });
@@ -941,7 +942,7 @@ auto Participant<SilKitConnectionT>::GetSystemController() -> Experimental::Serv
         config.name = SilKit::Core::Discovery::controllerTypeSystemController;
         config.network = "default";
 
-        return CreateController<Orchestration::SystemController>(config, std::move(supplementalData), true);
+        return CreateController<Orchestration::SystemController>(config, std::move(supplementalData), true, true);
     }
     return controller;
 }
@@ -1562,11 +1563,12 @@ template <class SilKitConnectionT>
 template <class ControllerT, typename... Arg>
 auto Participant<SilKitConnectionT>::CreateController(const SilKitServiceTraitConfigType_t<ControllerT>& config,
                                                       const SilKit::Core::SupplementalData& supplementalData,
-                                                      const bool publishService, Arg&&... arg) -> ControllerT*
+                                                      const bool publishServiceDiscovery,
+                                                      const bool registerSilKitService, Arg&&... arg) -> ControllerT*
 {
     SILKIT_ASSERT(config.network.has_value());
-    return CreateController<ControllerT>(config, *config.network, supplementalData, publishService,
-                                         std::forward<Arg>(arg)...);
+    return CreateController<ControllerT>(config, *config.network, supplementalData, publishServiceDiscovery,
+                                         registerSilKitService, std::forward<Arg>(arg)...);
 }
 
 template <class SilKitConnectionT>
@@ -1574,7 +1576,8 @@ template <class ControllerT, typename... Arg>
 auto Participant<SilKitConnectionT>::CreateController(const SilKitServiceTraitConfigType_t<ControllerT>& config,
                                                       const std::string& network,
                                                       const SilKit::Core::SupplementalData& supplementalData,
-                                                      const bool publishService, Arg&&... arg) -> ControllerT*
+                                                      const bool publishServiceDiscovery,
+                                                      const bool registerSilKitService, Arg&&... arg) -> ControllerT*
 {
     const auto serviceType = SilKitServiceTraitServiceType<ControllerT>::GetServiceType();
     if (config.name == "")
@@ -1606,15 +1609,29 @@ auto Participant<SilKitConnectionT>::CreateController(const SilKitServiceTraitCo
 
     controller->SetServiceDescriptor(std::move(descriptor));
 
-    _connection.RegisterSilKitService(controllerPtr);
+    if (registerSilKitService)
+    {
+        _connection.RegisterSilKitService(controllerPtr);
+    }
+
     const auto qualifiedName = config.name;
     controllerMap[qualifiedName] = std::move(controller);
 
-    if (publishService)
+    if (publishServiceDiscovery)
     {
         GetServiceDiscovery()->NotifyServiceCreated(controllerPtr->GetServiceDescriptor());
     }
     return controllerPtr;
+}
+
+template <class SilKitConnectionT>
+void Participant<SilKitConnectionT>::RegisterTimeSyncService(
+    SilKit::Services::Orchestration::TimeSyncService* controllerPtr)
+{
+    if (controllerPtr)
+    {
+        _connection.RegisterSilKitService(controllerPtr);
+    }
 }
 
 template <class SilKitConnectionT>